requirements/base.txt
asgiref==3.9.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==45.0.5
Django==5.2.4
django-admin-menu==1.6
django-ckeditor-5==0.2.18
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-js-asset==3.1.2
django-oauth-toolkit==3.0.1
django-phone-field==1.8.1
django-rest-passwordreset==1.5.0
djangorestframework==3.16.0
drf-yasg==1.21.10
future==1.0.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jwcrypto==1.5.6
libsass==0.23.0
MarkupSafe==3.0.2
mysqlclient==2.2.7
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
pycparser==2.22
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
serializers==0.2.4
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
